+++ src/pages/profile/settings.tsx
@@ -1,6 +1,6 @@
 import { Layout } from "@/components/layout";
 import { useLoadingIndicator } from "@/components/loading/indicator";
-import { logToastError, toast } from "@/components/Toast";
+import { logToastError, toast } from "@/components/toast";
 import {
   Button,
   DialogOverlay,
