+++ src/pages/scan.tsx
@@ -1,6 +1,6 @@
 import { Layout } from "@/components/layout";
 import { scanTypeOffsetStore, scanTypeStore } from "@/components/layout/footer";
-import { logToastError, toast } from "@/components/Toast";
+import { logToastError, toast } from "@/components/toast";
 import { ImageInput } from "@/components/UI";
 import { useSwipe } from "@/hooks";
 import { tw } from "@/utils";
